pyproject.toml
[tool.poetry]
name = "kaggle-imaterialist2020-model"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "configs", from = "tf_tpu_models/official/detection" },
    { include = "dataloader", from = "tf_tpu_models/official/detection" },
    { include = "executor", from = "tf_tpu_models/official/detection" },
    { include = "modeling", from = "tf_tpu_models/official/detection" },
    { include = "utils", from = "tf_tpu_models/official/detection" },
    { include = "hyperparameters", from = "tf_tpu_models" },
    { include = "kaggle_imaterialist2020_model" },
]

[tool.poetry.dependencies]
# Tensorflow 1.15.5 can't be installed with Python 3.8 or higher.
# https://github.com/tensorflow/tensorflow/issues/34302
python = ">=3.7.1,<3.8"
Cython = "^0.29.23"
matplotlib = "^3.4.1"
opencv-python-headless = "^4.5.1"
PyYAML = "^5.4.1"
Pillow = "^8.2.0"
scikit-learn = "^0.24.1"
pandas = "^1.2.4"
tqdm = "^4.60.0"
google-api-python-client = "^2.2.0"
oauth2client = "^4.1.3"
gcsfs = "^0.8.0"
spotty = "^1.3.2"
pycocotools = "^2.0.2"
absl-py = "^0.12.0"
segmentation = {git = "https://github.com/hrsma2i/segmentation.git", rev = "1c3972d80af0566cea8ef47d25a08de7ccb42222"}
typer = "^0.3.2"
# This CLI can't work with numpy 1.18.5 or higher.
numpy = "<=1.18.5"
typing-extensions = "^3.10.0"
google-cloud-bigquery = "^2.26.0"
tensorflow = "^2.7.0"
tensorflow-addons = "^0.15.0"
python-json-logger = "^2.0.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
mypy = "^0.812"
black = "^20.8b1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
